refactor: deduplicate code and improve file organization
No behavioral changes — purely structural cleanup:

- Extract callWithRecovery + isTokenError helpers (replaces 20+ scattered
  token-error-check → recoverToken → retry blocks across 6 files)
- Extract deriveFileKeys + newCTRStream (deduplicates 4 identical HKDF/AES-CTR
  routines in media.go)
- Extract mediaUnavailablePlaceholder + decryptMediaData (deduplicates 4
  placeholder message blocks and two-phase media decrypt pattern)
- Remove dead min() function (Go 1.21+ has builtin)
- Replace log.Printf with zerolog in pkg/line/client.go (5 calls)
- Split handle_message.go → handle_media.go (media helpers)
- Split sync.go → operations.go (operation dispatch)
