+++ pkg/connector/handle_message.go
@@ -169,64 +169,25 @@ func (lc *LineClient) queueIncomingMessage(msg *line.Message, opType int) {
 
 				if oid != "" {
 					var imgData []byte
-					var err error
-					if isPlainMedia {
-						imgData, err = client.DownloadOBSWithSID(oid, data.ID, "m")
-					} else {
-						imgData, err = client.DownloadOBS(oid, data.ID)
-					}
-
-					// Refresh token if we get a 401
-					if err != nil && (strings.Contains(err.Error(), "401") || lc.isRefreshRequired(err) || lc.isLoggedOut(err)) {
-						if errRecover := lc.recoverToken(ctx); errRecover == nil {
-							client = line.NewClient(lc.AccessToken)
-							if isPlainMedia {
-								imgData, err = client.DownloadOBSWithSID(oid, data.ID, "m")
-							} else {
-								imgData, err = client.DownloadOBS(oid, data.ID)
-							}
+					_, err := lc.callWithRecovery(ctx, func(c *line.Client) error {
+						var e error
+						if isPlainMedia {
+							imgData, e = c.DownloadOBSWithSID(oid, data.ID, "m")
 						} else {
-							lc.UserLogin.Bridge.Log.Warn().Err(errRecover).Msg("Failed to recover token for OBS download")
+							imgData, e = c.DownloadOBS(oid, data.ID)
 						}
-					}
+						return e
+					})
 
 					if err != nil {
-						lc.UserLogin.Bridge.Log.Warn().
-							Err(err).
-							Str("oid", oid).
-							Str("msg_id", data.ID).
-							Bool("plain_media", isPlainMedia).
+						lc.UserLogin.Bridge.Log.Warn().Err(err).Str("oid", oid).Str("msg_id", data.ID).Bool("plain_media", isPlainMedia).
 							Msg("Failed to download image from OBS, sending placeholder")
-						return &bridgev2.ConvertedMessage{
-							Parts: []*bridgev2.ConvertedMessagePart{
-								{
-									Type: event.EventMessage,
-									Content: &event.MessageEventContent{
-										MsgType:   event.MsgNotice,
-										Body:      "[Image unavailable — LINE media expired before it could be bridged]",
-										RelatesTo: replyRelatesTo,
-									},
-								},
-							},
-						}, nil
+						return mediaUnavailablePlaceholder("Image", replyRelatesTo), nil
 					}
 
-					// Decrypt image if it has keyMaterial (E2EE)
-					if decryptedBody != "" && strings.Contains(decryptedBody, "keyMaterial") {
-						var decryptInfo struct {
-							KeyMaterial string `json:"keyMaterial"`
-							FileName    string `json:"fileName"`
-						}
-						if err := json.Unmarshal([]byte(decryptedBody), &decryptInfo); err == nil && decryptInfo.KeyMaterial != "" {
-							decryptedImg, err := lc.decryptImageData(imgData, decryptInfo.KeyMaterial)
-							if err != nil {
-								lc.UserLogin.Bridge.Log.Error().
-									Err(err).
-									Msg("Failed to decrypt image data")
-								return nil, fmt.Errorf("failed to decrypt image data: %w", err)
-							}
-							imgData = decryptedImg
-						}
+					imgData, err = lc.decryptMediaData(imgData, decryptedBody, data.ContentMetadata)
+					if err != nil {
+						return nil, fmt.Errorf("failed to decrypt image data: %w", err)
 					}
 
 					// Upload to Matrix
@@ -282,79 +243,22 @@ func (lc *LineClient) queueIncomingMessage(msg *line.Message, opType int) {
 					if isPlainMedia {
 						sid = "m"
 					}
-					videoData, err := client.DownloadOBSWithSID(oid, data.ID, sid)
-
-					if err != nil && (strings.Contains(err.Error(), "401") || lc.isRefreshRequired(err) || lc.isLoggedOut(err)) {
-						if errRecover := lc.recoverToken(ctx); errRecover == nil {
-							client = line.NewClient(lc.AccessToken)
-							videoData, err = client.DownloadOBSWithSID(oid, data.ID, sid)
-						} else {
-							lc.UserLogin.Bridge.Log.Warn().Err(errRecover).Msg("Failed to recover token for OBS download")
-						}
-					}
+					var videoData []byte
+					_, err := lc.callWithRecovery(ctx, func(c *line.Client) error {
+						var e error
+						videoData, e = c.DownloadOBSWithSID(oid, data.ID, sid)
+						return e
+					})
 
 					if err != nil {
-						lc.UserLogin.Bridge.Log.Warn().
-							Err(err).
-							Str("oid", oid).
-							Str("msg_id", data.ID).
-							Bool("plain_media", isPlainMedia).
+						lc.UserLogin.Bridge.Log.Warn().Err(err).Str("oid", oid).Str("msg_id", data.ID).Bool("plain_media", isPlainMedia).
 							Msg("Failed to download video from OBS, sending placeholder")
-						return &bridgev2.ConvertedMessage{
-							Parts: []*bridgev2.ConvertedMessagePart{
-								{
-									Type: event.EventMessage,
-									Content: &event.MessageEventContent{
-										MsgType:   event.MsgNotice,
-										Body:      "[Video unavailable — LINE media expired before it could be bridged]",
-										RelatesTo: replyRelatesTo,
-									},
-								},
-							},
-						}, nil
+						return mediaUnavailablePlaceholder("Video", replyRelatesTo), nil
 					}
 
-					if decryptedBody != "" && strings.Contains(decryptedBody, "keyMaterial") {
-						var decryptInfo struct {
-							KeyMaterial string `json:"keyMaterial"`
-							FileName    string `json:"fileName"`
-						}
-						if err := json.Unmarshal([]byte(decryptedBody), &decryptInfo); err == nil && decryptInfo.KeyMaterial != "" {
-							lc.UserLogin.Bridge.Log.Debug().
-								Str("key_material_len", fmt.Sprintf("%d", len(decryptInfo.KeyMaterial))).
-								Str("file_name", decryptInfo.FileName).
-								Msg("Decrypting E2EE video")
-
-							decryptedVideo, err := lc.decryptImageData(videoData, decryptInfo.KeyMaterial)
-							if err != nil {
-								lc.UserLogin.Bridge.Log.Error().
-									Err(err).
-									Msg("Failed to decrypt video data")
-								return nil, fmt.Errorf("failed to decrypt video data: %w", err)
-							}
-							videoData = decryptedVideo
-							lc.UserLogin.Bridge.Log.Info().
-								Int("decrypted_size", len(videoData)).
-								Msg("Successfully decrypted video")
-						}
-					}
-
-					if encKM := data.ContentMetadata["ENC_KM"]; encKM != "" && len(videoData) > 32 {
-						lc.UserLogin.Bridge.Log.Debug().
-							Str("enc_km_preview", encKM[:min(20, len(encKM))]+"...").
-							Msg("Decrypting video using ENC_KM from metadata")
-
-						decryptedVideo, err := lc.decryptImageData(videoData, encKM)
-						if err != nil {
-							lc.UserLogin.Bridge.Log.Error().
-								Err(err).
-								Msg("Failed to decrypt video data from ENC_KM")
-							return nil, fmt.Errorf("failed to decrypt video data: %w", err)
-						}
-						videoData = decryptedVideo
-						lc.UserLogin.Bridge.Log.Info().
-							Int("decrypted_size", len(videoData)).
-							Msg("Successfully decrypted video from ENC_KM")
+					videoData, err = lc.decryptMediaData(videoData, decryptedBody, data.ContentMetadata)
+					if err != nil {
+						return nil, fmt.Errorf("failed to decrypt video data: %w", err)
 					}
 
 					fileName := data.ContentMetadata["FILE_NAME"]
@@ -444,25 +348,16 @@ func (lc *LineClient) queueIncomingMessage(msg *line.Message, opType int) {
 					if isPlainMedia {
 						sid = "m"
 					}
-					fileData, err := client.DownloadOBSWithSID(oid, data.ID, sid)
+					var fileData []byte
+					_, err := lc.callWithRecovery(ctx, func(c *line.Client) error {
+						var e error
+						fileData, e = c.DownloadOBSWithSID(oid, data.ID, sid)
+						return e
+					})
 					if err != nil {
-						lc.UserLogin.Bridge.Log.Warn().
-							Err(err).
-							Str("oid", oid).
-							Bool("plain_media", isPlainMedia).
+						lc.UserLogin.Bridge.Log.Warn().Err(err).Str("oid", oid).Bool("plain_media", isPlainMedia).
 							Msg("Failed to download file from OBS, sending placeholder")
-						return &bridgev2.ConvertedMessage{
-							Parts: []*bridgev2.ConvertedMessagePart{
-								{
-									Type: event.EventMessage,
-									Content: &event.MessageEventContent{
-										MsgType:   event.MsgNotice,
-										Body:      "[File unavailable — LINE media expired before it could be bridged]",
-										RelatesTo: replyRelatesTo,
-									},
-								},
-							},
-						}, nil
+						return mediaUnavailablePlaceholder("File", replyRelatesTo), nil
 					}
 
 					// Try to decrypt using keyMaterial from encrypted payload
@@ -571,56 +466,22 @@ func (lc *LineClient) queueIncomingMessage(msg *line.Message, opType int) {
 					if isPlainMedia {
 						sid = "m"
 					}
-					audioData, err := client.DownloadOBSWithSID(oid, data.ID, sid)
-
-					if err != nil && (strings.Contains(err.Error(), "401") || lc.isRefreshRequired(err) || lc.isLoggedOut(err)) {
-						if errRecover := lc.recoverToken(ctx); errRecover == nil {
-							client = line.NewClient(lc.AccessToken)
-							audioData, err = client.DownloadOBSWithSID(oid, data.ID, sid)
-						}
-					}
+					var audioData []byte
+					_, err := lc.callWithRecovery(ctx, func(c *line.Client) error {
+						var e error
+						audioData, e = c.DownloadOBSWithSID(oid, data.ID, sid)
+						return e
+					})
 
 					if err != nil {
-						lc.UserLogin.Bridge.Log.Warn().
-							Err(err).
-							Str("oid", oid).
-							Str("msg_id", data.ID).
-							Bool("plain_media", isPlainMedia).
+						lc.UserLogin.Bridge.Log.Warn().Err(err).Str("oid", oid).Str("msg_id", data.ID).Bool("plain_media", isPlainMedia).
 							Msg("Failed to download audio from OBS, sending placeholder")
-						return &bridgev2.ConvertedMessage{
-							Parts: []*bridgev2.ConvertedMessagePart{
-								{
-									Type: event.EventMessage,
-									Content: &event.MessageEventContent{
-										MsgType:   event.MsgNotice,
-										Body:      "[Audio unavailable — LINE media expired before it could be bridged]",
-										RelatesTo: replyRelatesTo,
-									},
-								},
-							},
-						}, nil
+						return mediaUnavailablePlaceholder("Audio", replyRelatesTo), nil
 					}
 
-					// Decrypt audio if it has keyMaterial (E2EE)
-					if decryptedBody != "" && strings.Contains(decryptedBody, "keyMaterial") {
-						var decryptInfo struct {
-							KeyMaterial string `json:"keyMaterial"`
-						}
-						if err := json.Unmarshal([]byte(decryptedBody), &decryptInfo); err == nil && decryptInfo.KeyMaterial != "" {
-							decryptedAudio, err := lc.decryptImageData(audioData, decryptInfo.KeyMaterial)
-							if err != nil {
-								return nil, fmt.Errorf("failed to decrypt audio data: %w", err)
-							}
-							audioData = decryptedAudio
-						}
-					}
-
-					if encKM := data.ContentMetadata["ENC_KM"]; encKM != "" && len(audioData) > 32 {
-						decryptedAudio, err := lc.decryptImageData(audioData, encKM)
-						if err != nil {
-							return nil, fmt.Errorf("failed to decrypt audio data: %w", err)
-						}
-						audioData = decryptedAudio
+					audioData, err = lc.decryptMediaData(audioData, decryptedBody, data.ContentMetadata)
+					if err != nil {
+						return nil, fmt.Errorf("failed to decrypt audio data: %w", err)
 					}
 
 					var duration int
@@ -789,36 +650,4 @@ func (lc *LineClient) queueIncomingMessage(msg *line.Message, opType int) {
 			}, nil
 		},
 	})
-}
-
-// resolveReplyRelatesTo looks up the Matrix event ID for a replied-to LINE message.
-func (lc *LineClient) resolveReplyRelatesTo(ctx context.Context, data *line.Message) *event.RelatesTo {
-	if data == nil {
-		return nil
-	}
-
-	relatedID := data.RelatedMessageID
-	if relatedID == "" && data.ContentMetadata != nil {
-		relatedID = data.ContentMetadata["message_relation_server_message_id"]
-	}
-
-	if relatedID == "" {
-		return nil
-	}
-
-	if data.MessageRelationType != 0 && data.MessageRelationType != 3 {
-		return nil
-	}
-
-	dbMsg, err := lc.UserLogin.Bridge.DB.Message.GetPartByID(ctx, lc.UserLogin.ID, networkid.MessageID(relatedID), "")
-	if err != nil {
-		lc.UserLogin.Bridge.Log.Debug().Err(err).Str("related_msg_id", relatedID).Msg("Failed to lookup reply target")
-		return nil
-	}
-	if dbMsg == nil || dbMsg.MXID == "" {
-		lc.UserLogin.Bridge.Log.Debug().Str("related_msg_id", relatedID).Msg("No Matrix event found for reply target")
-		return nil
-	}
-
-	return &event.RelatesTo{InReplyTo: &event.InReplyTo{EventID: dbMsg.MXID}}
 }
